Инсталл виа хсерч фикс

+++ heroku/modules/HSearch.py
@@ -601,17 +601,33 @@ class HSearch(loader.Module):
         if not link.startswith("https://api.fixyres.com/module/"):
             return
         
-        try:
-            await self.lookup("loader").download_and_install(link, None)
-            
-            await asyncio.sleep(1)
+        if self.lookup("FHeta"):
+            return
 
-            if self.lookup("loader").fully_loaded:
-                self.lookup("loader").update_modules_in_db()
+        try:
+            parsed_url = urlparse(link)
+            qp = parse_qs(parsed_url.query)
             
-            rose_msg = await message.respond("🌹")
-            await asyncio.sleep(1)
-            await rose_msg.delete()
-            await message.delete()
+            ohd = qp.get('ohd', ['False'])[0]
+            
+            if ohd.lower() == 'false':
+                cross_msg = await message.respond("❌")
+                await asyncio.sleep(1)
+                await cross_msg.delete()
+                await message.delete()
+                return
+            
+            if ohd.lower() == 'true':
+                await self.lookup("loader").download_and_install(link, None)
+                
+                await asyncio.sleep(1)
+
+                if self.lookup("loader").fully_loaded:
+                    self.lookup("loader").update_modules_in_db()
+                
+                rose_msg = await message.respond("🌹")
+                await asyncio.sleep(1)
+                await rose_msg.delete()
+                await message.delete()
         except:
             pass
